+++ plugins/catalog-react/src/components/EntityRefLink/humanize.test.ts
@@ -34,6 +34,23 @@ describe('humanizeEntityRef', () => {
     expect(title).toEqual('component:software');
   });
 
+  it('formats entity in default namespace without skipping default namespace', () => {
+    const entity = {
+      apiVersion: 'v1',
+      kind: 'Component',
+      metadata: {
+        name: 'software',
+      },
+      spec: {
+        owner: 'guest',
+        type: 'service',
+        lifecycle: 'production',
+      },
+    };
+    const title = humanizeEntityRef(entity, { defaultNamespace: false });
+    expect(title).toEqual('component:default/software');
+  });
+
   it('formats entity in other namespace', () => {
     const entity = {
       apiVersion: 'v1',
@@ -50,6 +67,24 @@ describe('humanizeEntityRef', () => {
     };
     const title = humanizeEntityRef(entity);
     expect(title).toEqual('component:test/software');
+  });
+
+  it('formats entity in other namespace and hides this namespace', () => {
+    const entity = {
+      apiVersion: 'v1',
+      kind: 'Component',
+      metadata: {
+        name: 'software',
+        namespace: 'test',
+      },
+      spec: {
+        owner: 'guest',
+        type: 'service',
+        lifecycle: 'production',
+      },
+    };
+    const title = humanizeEntityRef(entity, { defaultNamespace: 'test' });
+    expect(title).toEqual('component:software');
   });
 
   it('formats entity and hides default kind', () => {
@@ -70,6 +105,27 @@ describe('humanizeEntityRef', () => {
     expect(title).toEqual('test/software');
   });
 
+  it('formats entity and hides default kind and hiding namespace', () => {
+    const entity = {
+      apiVersion: 'v1',
+      kind: 'Component',
+      metadata: {
+        name: 'software',
+        namespace: 'test',
+      },
+      spec: {
+        owner: 'guest',
+        type: 'service',
+        lifecycle: 'production',
+      },
+    };
+    const title = humanizeEntityRef(entity, {
+      defaultKind: 'Component',
+      defaultNamespace: 'test',
+    });
+    expect(title).toEqual('software');
+  });
+
   it('formats entity name in default namespace', () => {
     const entityName = {
       kind: 'Component',
@@ -78,6 +134,16 @@ describe('humanizeEntityRef', () => {
     };
     const title = humanizeEntityRef(entityName);
     expect(title).toEqual('component:software');
+  });
+
+  it('formats entity name in default namespace and does not skip default namespace', () => {
+    const entityName = {
+      kind: 'Component',
+      namespace: 'default',
+      name: 'software',
+    };
+    const title = humanizeEntityRef(entityName, { defaultNamespace: false });
+    expect(title).toEqual('component:default/software');
   });
 
   it('formats entity name in other namespace', () => {
@@ -91,6 +157,19 @@ describe('humanizeEntityRef', () => {
     expect(title).toEqual('component:test/software');
   });
 
+  it('formats entity name in other namespace with skipping this namespace', () => {
+    const entityName = {
+      kind: 'Component',
+      namespace: 'test',
+      name: 'software',
+    };
+
+    const title = humanizeEntityRef(entityName, {
+      defaultNamespace: 'test',
+    });
+    expect(title).toEqual('component:software');
+  });
+
   it('renders link for entity name and hides default kind', () => {
     const entityName = {
       kind: 'Component',
@@ -102,5 +181,32 @@ describe('humanizeEntityRef', () => {
       defaultKind: 'component',
     });
     expect(title).toEqual('test/software');
+  });
+
+  it('renders link for entity name and hides default kind with skipping namespace', () => {
+    const entityName = {
+      kind: 'Component',
+      namespace: 'test',
+      name: 'software',
+    };
+
+    const title = humanizeEntityRef(entityName, {
+      defaultKind: 'component',
+      defaultNamespace: 'test',
+    });
+    expect(title).toEqual('software');
+  });
+
+  it('formats entity name in default namespace without skip of default namespace', () => {
+    const entityName = {
+      kind: 'Component',
+      namespace: 'default',
+      name: 'software',
+    };
+
+    const title = humanizeEntityRef(entityName, {
+      defaultNamespace: false,
+    });
+    expect(title).toEqual('component:default/software');
   });
 });