+++ plugins/auth-node/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-auth-node
 
+## 0.2.5-next.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/config@1.0.2-next.0
+  - @backstage/errors@1.1.1-next.0
+  - @backstage/backend-common@0.15.1-next.3
+
 ## 0.2.5-next.2
 
 ### Patch Changes