+++ plugins/kubernetes/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @backstage/plugin-kubernetes
 
+## 0.7.2-next.3
+
+### Patch Changes
+
+- 7d47def9c4: Removed dependency on `@types/jest`.
+- 19a27929fb: Reset error state on success
+- Updated dependencies
+  - @backstage/plugin-catalog-react@1.1.4-next.2
+  - @backstage/catalog-model@1.1.1-next.0
+  - @backstage/config@1.0.2-next.0
+  - @backstage/core-components@0.11.1-next.3
+  - @backstage/core-plugin-api@1.0.6-next.3
+  - @backstage/plugin-kubernetes-common@0.4.2-next.1
+
 ## 0.7.2-next.2
 
 ### Patch Changes